+++ luminex/src/org/labkey/luminex/query/LuminexProtocolSchema.java
@@ -353,10 +353,10 @@ public class LuminexProtocolSchema extends AssayProtocolSchema
     }
 
     @Override
-    public LuminexDataTable createDataTable(ContainerFilter cf, boolean includeCopiedToStudyColumns)
+    public LuminexDataTable createDataTable(ContainerFilter cf, boolean includeLinkedToStudyColumns)
     {
         LuminexDataTable table = new LuminexDataTable(this, cf);
-        if (includeCopiedToStudyColumns)
+        if (includeLinkedToStudyColumns)
         {
             addLinkedToStudyColumns(table, true);
         }
