+++ viability/src/org/labkey/viability/ViabilityAssaySchema.java
@@ -120,7 +120,7 @@ public class ViabilityAssaySchema extends AssayProtocolSchema
     public FilteredTable createDataTable(ContainerFilter cf, boolean includeLinkedToStudyColumns)
     {
         // UNDONE: add link to study columns when link to study is implemented
-        //addCopiedToStudyColumns(table, protocol, schema.getUser(), "rowId", true);
+        //addLinkedToStudyColumns(table, protocol, schema.getUser(), "rowId", true);
         return new ResultsTable(cf);
     }
